Re-add hiding of certain nodes in the CST.

+++ lib/endb_lib/endb.h
@@ -231,16 +231,13 @@ typedef void (*endb_arrow_array_stream_consumer_on_init_stream_callback)(struct 
 
 typedef void (*endb_arrow_array_stream_consumer_on_success_callback)(const uint8_t*, uintptr_t);
 
-typedef void (*endb_parse_sql_cst_on_open_callback)(const uint8_t*, uintptr_t);
+typedef void (*endb_parse_sql_cst_on_open_callback)(const uint8_t*, uint32_t);
 
 typedef void (*endb_parse_sql_cst_on_close_callback)(void);
 
-typedef void (*endb_parse_sql_cst_on_literal_callback)(const uint8_t*,
-                                                       uintptr_t,
-                                                       uintptr_t,
-                                                       uintptr_t);
+typedef void (*endb_parse_sql_cst_on_literal_callback)(const uint8_t*, uint32_t, uint32_t, uint32_t);
 
-typedef void (*endb_parse_sql_cst_on_pattern_callback)(uintptr_t, uintptr_t);
+typedef void (*endb_parse_sql_cst_on_pattern_callback)(uint32_t, uint32_t);
 
 typedef void (*endb_render_json_error_report_on_success_callback)(const char*);
